requirements.txt
anyio==3.6.2
Cerberus==1.3.4
click==8.1.3
fastapi==0.88.0
h11==0.14.0
idna==3.4
psycopg2==2.9.5
psycopg2-binary==2.9.5
pydantic==1.10.2
python-slugify==7.0.0
sniffio==1.3.0
starlette==0.22.0
text-unidecode==1.3
typing_extensions==4.4.0
uvicorn==0.20.0
python-dotenv==0.21.0
SQLAlchemy~=1.4.45